fix: solve room service issue and other dependencies of the project to build it correctly. Actually it's running in port 8080. It has been deleted a duplicated websockets file

+++ src/main/kotlin/edu/eci/arsw/code_arena/websocket/GameWebSocketController.kt
@@ -184,7 +184,7 @@ class GameWebSocketController(
             val answerResult = AnswerResultMessage(
                 isCorrect = playerAnswer.isCorrect,
                 score = playerAnswer.score,
-                responseTime = playerAnswer.responseTime,
+                responseTime = playerAnswer.responseTime.toLong(),
                 message = if (playerAnswer.isCorrect) "Correct!" else "Incorrect!"
             )
